requirements.txt
# GraphRAG Multi-Agent 项目依赖

# 核心框架
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# LLM 支持
langchain-anthropic>=0.2.0  # DeepSeek API 支持
langchain-ollama>=0.2.0    # Ollama 支持

# 向量数据库
chromadb>=0.5.0

# 向量嵌入
sentence-transformers>=3.0.0
transformers>=4.40.0
torch>=2.0.0

# 图数据库
networkx>=3.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 工具库
tqdm>=4.60.0
python-dotenv>=1.0.0

# 开发工具
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# 可选：用于更好的性能
# faiss-cpu>=1.7.0  # 如果使用 FAISS 作为向量索引
# sentencepiece>=0.1.99  # 某些模型需要